requirements.txt
Flask==3.0.3
flask_cors==5.0.0
Werkzeug==3.0.6
gunicorn==23.0.0
bcrypt==4.1.3
boto3==1.34.162
pandas==2.2.3
openpyxl==3.1.5
requests==2.31.0
python-dotenv==1.0.1
google-generativeai==0.5.4
langchain==0.1.20
langchain-openai==0.1.7
langchain-google-genai==1.0.4
chromadb==0.5.23
flashrank==0.2.9
websocket-client
openai==1.29.0
beautifulsoup4==4.12.3
anthropic~=0.23.1
tiktoken==0.7.0
python-docx==1.1.0
